+++ opm/parser/eclipse/EclipseState/Grid/FieldPropsManager.hpp
@@ -22,6 +22,7 @@
 #include <memory>
 #include <vector>
 #include <opm/parser/eclipse/EclipseState/Grid/TranCalculator.hpp>
+#include <opm/parser/eclipse/EclipseState/Grid/FieldData.hpp>
 
 namespace Opm {
